Fix overshadowing of events variable.
Consider an instance of the `Attr` class
constructed with a node and an event attribute,
such as `onkeyup`. Previously, calling `updateElement` on this instance
inserted the DOM event object into the local mesh with the name `event`.
Even after the event callback finished, the DOM event object remained in the
local mesh. As the local mesh has priority, the variable name event
could not be used in other places.

This commit makes two changes to fix this bug.
* Give the DOM event object in the local mesh a more specific name. (`domElement`)
* Return domElement to its initial state after the event callback
  finishes execution.

Adding a test, `tests/handlers/attr.js`, although the test
does not entirely replicate the error. If this
bug continues to be troublesome, this test is a good
starting point.